requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
requests>=2.28.0
pyyaml>=6.0.0
torch>=2.0.0
tqdm>=4.65.0
